% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{annots_to_capt}
\alias{annots_to_capt}
\title{Convert annotations to a capture history}
\usage{
annots_to_capt(annots, gap = 0)
}
\arguments{
\item{annots}{A dataframe containing at least \code{tstart} and \code{tend} columns,
and a \code{rec} column with recorder IDs. The timestamp columns are typically
numeric or POSIXt. AviaNZ annotations can be read into
this form directly with \code{\link[=readAnnots]{readAnnots()}} or \code{\link[=readAnnotsFile]{readAnnotsFile()}}.}

\item{gap}{A number in the same units as \verb{tstart,tend}.
Annotations that are separated by this much or more will be merged into
a single call. Default 0 merges any annotations that overlap or touch;
setting it larger is useful if the calls are often detected as multiple
annotations which should be merged into a single unit. Setting it below 0
will only merge annotations that overlap by at least \code{gap} time units, which
could help avoid falsely merging distinct calls if they occur frequently.}
}
\value{
A dataframe where... ??? TODO describe exactly the columns
Recaptures of the same call will be assigned the same ID.
}
\description{
Converts a dataframe of timestamped annotations into a shape compatible
with \code{\link[ascr:create.capt]{ascr::create.capt()}}.
}
\details{
Acoustic SCR requires identifying which detections in a survey represent
re-captures of the same cue. This function does that simply by using
temporal overlap: any annotations that overlap in time by at least \code{gap}
seconds are assumed to be recaptures of the same call, and assigned a
shared call ID. Note that this will merge both recaptures across recorders,
and also annotations within each recorder, such as when several annotations
are produced for the same call.
Note that any further annotation details (call type, species) will be ignored,
i.e. calls labelled as different species, overlapping in time, will still be
merged. If you wish to use such information, split the data by the groups yourself
and apply this function to each separately.
}
\examples{
## rec1 captures what looks like two pieces of the same call
## rec2 recaptures part of the same call, and another cue:
annots <- data.frame(rec=c("rec1", "rec1", "rec2", "rec2"),
  tstart=c(0.0, 5.5, 1.5, 9), tend=c(5.0, 6.5, 6.0, 13))
## merge into two unique calls + 1 recapture:
annots_to_capt(annots, gap=0)
## or we can separate out the short annotation in rec1 as another call:
annots_to_capt(annots, gap=-1)
}
