% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avianz2r-package.R
\docType{package}
\name{avianz2r-package}
\alias{avianz2r}
\alias{avianz2r-package}
\title{avianz2r: Reading and Processing of AviaNZ Annotations in R}
\description{
Functions for reading AviaNZ-format annotations into R, doing simple processing, and preparing them for acoustic SCR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://avianz.net/}
  \item Report bugs at \url{https://github.com/jjuod/avianz2r/issues}
}

}
\author{
\strong{Maintainer}: Julius Juodakis \email{j.juodakis@gmail.com} (\href{https://orcid.org/0000-0001-5600-974X}{ORCID})

}
\keyword{internal}
