% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lag_clocks}
\alias{lag_clocks}
\title{Shift timestamps for each recorder}
\usage{
lag_clocks(data, lags)
}
\arguments{
\item{data}{A dataframe containing at least \code{tstart} and \code{tend} columns with
values to be adjusted, and a \code{rec} with recorder IDs for matching with lags.
AviaNZ annotations can be read into this form directly with \code{\link[=readAnnots]{readAnnots()}} or
\code{\link[=readAnnotsFile]{readAnnotsFile()}}.}

\item{lags}{A dataframe containing two columns: \code{rec} with recorder IDs,
and \code{lag} with the lags to be applied to the each recorder. Recorders will be matched
with \code{data}, and \code{lag} simply added to the \code{tstart} and \code{tend} of each annotation.
Lags can be positive, zero or negative (i.e. leads) numbers, provided on the same
scale as the time in \code{tstart} or \code{tend}, so typically seconds.
Lags must be provided for all recorders present in data (provide 0 if some
recorders do not need adjustment).}
}
\value{
The input dataframe with \code{tstart} and \code{tend} values adjusted.
}
\description{
Applies recorder-level adjustments (lag or lead) to synchronize timestamps across recorders.
}
\examples{
# Read annotations from two recorders
annotdir = system.file("extdata", package="avianz2r", mustWork=TRUE)
annots <- readAnnots(annotdir)
# Adjust the second recorder by +10 s. The third recorder is not present in data
clocklags = data.frame(rec=c("recA", "recB", "recC"),
                       lag=c(0, 10, -5))
lag_clocks(annots, clocklags)
}
