% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{merge_syllables}
\alias{merge_syllables}
\title{Merge timestamps with small gaps}
\usage{
merge_syllables(annots, gap)
}
\arguments{
\item{annots}{A dataframe containing at least \code{tstart} and \code{tend} columns,
and a \code{rec} column with recorder IDs. The timestamp columns are typically
numeric or POSIXt. AviaNZ annotations can be read into this form directly
with \code{\link[=readAnnots]{readAnnots()}} or \code{\link[=readAnnotsFile]{readAnnotsFile()}}. Any additional columns will be
ignored -- edit this function to add additional merging rules if you wish.}

\item{gap}{A number in the same units as \verb{tstart,tend}. Annotations that are
separated by this much or less will be merged into a single call. I.e. 0
merges any annotations that overlap or touch; positive \code{gap} is likely the most
useful for merging syllables, but it can also be set below 0 to only
merge annotations that overlap by at least \code{gap} time units.}
}
\value{
A dataframe containing \code{tstart} and \code{tend} columns corresponding to the
timestamps after merging, and \code{rec} with recorder ID as provided in the input.
}
\description{
Merges annotations separated by at most a chosen time gap within each
recorder.
}
\details{
This is a replacement for \code{\link[=annots_to_calls]{annots_to_calls()}}, in case you want to only merge
subsequent annotations within each recorder and not across groups, e.g. for
combining syllables into calls. This is useful if (1) you want a more liberal
gap to be applied for within-recorder merges than for between-recorder
merges, (2) preserving additional call information is needed -- this function
can be modified more easily than \code{\link[=annots_to_calls]{annots_to_calls()}}.
}
\examples{
## several detections over 2 recorders:
annots <- data.frame(rec=c("rec1", "rec1", "rec1", "rec2", "rec2"),
  tstart=c(0.0, 3.5, 7.1, 1.5, 5.5), tend=c(3.0, 4.5, 8.1, 6.0, 8.0))
## merge into 3 unique calls:
merge_syllables(annots, gap=2)
}
