% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{plot_annots}
\alias{plot_annots}
\title{Visualize annotation times}
\usage{
plot_annots(annots, days = NULL, hours = NULL)
}
\arguments{
\item{annots}{A dataframe containing at least \code{tstart} and \code{tend} columns.
The annotations will be shown as line segments covering \verb{[tstart; tend]}.
These columns must be POSIXt timestamps.
Columns \code{rec} and/or \code{species}, if present and containing more than one unique value,
will be used to spread out the timestamps along the Y axis. AviaNZ annotations
can be read into this form directly with \code{\link[=readAnnots]{readAnnots()}} or \code{\link[=readAnnotsFile]{readAnnotsFile()}}.}

\item{days, hours}{Optional, numeric vectors. If specified, will only plot annotations
starting within this(these) day(s) of the month or hour(s).
Typically, only several hours of data can be usefully shown on this type of plot,
as afterwards calls become too short and dense to distinguish. These parameters
are provided as a quick way to inspect the data in parts.}
}
\value{
A ggplot plot object. Such plots can be modified with the \code{+} operator;
see \link{ggplot2} documentation for details.
}
\description{
Plots the provided timestamps for a quick visual overview, separating by species and recorders.
}
\examples{
annotdir = system.file("extdata", package="avianz2r", mustWork=TRUE)
df <- readAnnots(annotdir)
plot_annots(df, hours=1)
# modifying the plot to split off custom time periods as rows:
library(ggplot2)
library(lubridate)
plot_annots(df[df$rec=="recA",]) +
  facet_wrap(~minute(tstart), nrow=2, scales="free") +
  scale_x_datetime(date_breaks="10 sec", date_minor_breaks="1 sec", date_label="\%M.\%S")
}
