% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{prepare_capt}
\alias{prepare_capt}
\title{Final formatting of call history for ascr}
\usage{
prepare_capt(calls, gpspos, survey.recs = NULL)
}
\arguments{
\item{calls}{A call history dataframe with columns \code{id} (arbitrary ID of each
call) and \code{rec} (recorder ID on which this capture/recapture was
registered). Auxiliary info may be provided in columns \code{bearing}, \code{dist},
\code{ss}, \code{toa}.}

\item{gpspos}{A dataframe with the recorder locations: columns \code{rec}, \code{east}
and \code{north} with recorder ID, eastings and northings. \code{rec} is used for
matching with \code{calls} and creating a new, numeric recorder ID. Coordinates
will be unchanged, so you should prepare them as for \code{\link[ascr:fit.ascr]{ascr::fit.ascr()}}. In
principle any system is suitable; we recommend using \code{\link[rgdal:spTransform-methods]{rgdal::spTransform()}}
to project the collected GPS data into UTM or similar projections, so that
the inputs are meters easting or northing, for meaningful interpretation of
the results.}

\item{survey.recs}{Optional: a vector of recorder IDs from \code{rec} to include.
Defaults to \code{NULL} = include all.
For example, it allows to conveniently exclude any recorders that did not
work on a particular occasion without editing the location data.}
}
\value{
A named list with \code{calls} and \code{traps} elements which can be directly
used in \code{\link[ascr:create.capt]{ascr::create.capt()}}. The main difference from the inputs is that
\code{calls} will have a numeric recorder ID, pointing to rows in \code{traps},
and dummy columns for session and occasion, while \code{traps} will be a matrix
without any recorder ID column.
}
\description{
Makes some minor formatting changes to allow using a call history in
\code{\link[ascr:create.capt]{ascr::create.capt()}}.
}
\examples{

# Recorder positions in eastings-northings:
gpspos = data.frame(rec=c("rec1", "rec2"),
                    east=c(-50, 50),
                    north=c(0, 175))

# Call history, from annotations:
annots =data.frame(rec=c("rec1", "rec1", "rec2", "rec2"),
  tstart=c(0.0, 5.5, 1.5, 9), tend=c(5.0, 6.5, 6.0, 13))
calls = annots_to_calls(annots)
l = prepare_capt(calls, gpspos)

# Prepare the actual capt object and run acoustic SCR
\dontrun{
library(ascr)
capt = ascr::create.capt(l$calls, traps=l$traps)
mask = ascr::create.mask(traps, buffer=700)
cr = ascr::fit.ascr(capt, traps, mask)
}
}
