% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readAnnots}
\alias{readAnnots}
\title{Read AviaNZ annotations}
\usage{
readAnnots(
  dir,
  time.formats = c("\%Y\%m\%d_\%H\%M\%S", "\%d\%m\%y_\%H\%M\%S"),
  exact = TRUE
)
}
\arguments{
\item{dir}{Directory path containing AviaNZ annotations ("new-style" JSON format, as
defined in AviaNZ documentation). Will be explored recursively. File tree and
naming must follow one of these formats:
\itemize{
\item \code{rec_date_time.wav.data} format. The filename of each \code{.data} file must contain
exactly three underscore-separated components, which will be parsed into
recorder ID, date and time. These files may be arranged into subdirectories in
anyway -- this arrangement will be ignored, and only the filename parsed.
\item \code{rec/date_time.wav.data} format. The filename of each \code{.data} file must contain
exactly two underscore-separated components, which will be parsed into date
and time. These files must be arranged into subdirectories. The recorder name
will be determined from the lowest-level subdirectory: so the recorder for
\verb{projectX/site1/day1/rec2/20010101_001234.wav.data} will be identified as \code{rec2}.
Higher-level directory structure, such as the \code{site1} here, are not important.
}}

\item{time.formats}{All permissible timestamp formats for the files, in \code{\link[=strptime]{strptime()}}
format specification.}

\item{exact}{If \code{FALSE}, a more sophisticated process of resolving timestamps will
be applied, which might help when a mixture of formats is present.
See \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} for details. Use with care, as timestamps
are critical for subsequent handling -- in case of any suspicious outputs,
a safer solution may be to separate the data into subdirectories by format,
and load each separately.}
}
\value{
A dataframe (namely, tibble) with one row per each species label, and
at least 5 columns:
\itemize{
\item \code{tstart} and \code{tend} (numeric, annotation position in seconds
from the file start)
\item \code{freqmin} and \code{freqmax} (numeric, frequency bounds in Hz,
or 0, if not specified)
\item any further columns from the label, typically \code{species},
\code{certainty} and \code{filter}
\item \code{ftime} (POSIXct, starting time of the file)
\item \code{rec} (name of the recorder).
}

Note that an AviaNZ annotation may have multiple labels, corresponding to
different species; these will be parsed into individual rows.
}
\description{
Reads one directory of AviaNZ format .data files into an R dataframe.
}
\examples{
annotdir = system.file("extdata", package="avianz2r", mustWork=TRUE)
df <- readAnnots(annotdir)
}
