% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readTestAnnots}
\alias{readTestAnnots}
\title{Read AviaNZ recognizer test results}
\usage{
readTestAnnots(dir, ...)
}
\arguments{
\item{dir}{Directory path containing AviaNZ annotations.}

\item{...}{Passed directly to \code{\link[=readAnnots]{readAnnots()}}.}
}
\value{
A dataframe (namely, tibble) with one row per each species label, all
columns as described in \code{\link[=readAnnots]{readAnnots()}}, plus \code{detector}:
"manual", "auto", or "auto_wv" indicating the original annotations, final
detections, or detections after the wavelet stage only, respectively. For
filters without a CNN, rows corresponding "auto" and "auto_wv" should be
the same.
}
\description{
Reads in data files produced in AviaNZ recognizer testing.
}
\details{
When testing a recognizer in AviaNZ, the automatic detections are stored in
.tmpdata files, .tmp2data files (automatic detections before CNN), and also
the original .data files with manual annotations might be of interest. This
function reads in all these sets of data. Arguments are passed to
\code{\link[=readAnnots]{readAnnots()}} - see there for details.
}
